server/requirements.txt
fire
flask
flask-cors
pydantic
mypy
mypy-extensions
flake8
pytest
numpy
opencv-python
Pillow
torch
torchvision
tqdm
timm
h5py
git+https://github.com/deepglint/unicom.git